src/gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-aiplatform[generative-ai]==1.74.0
# httpx pinned: 0.28+ breaks starlette TestClient (starlette 0.35 compat)
httpx==0.27.2
